costpilot/agents/seed/pyproject.toml
[project]
name = "costpilot-seed"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary>=2.9",
    "faker>=28.0",
]
